% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airshed.R
\name{airshed}
\alias{airshed}
\title{Create HYSPLIT-derived Seasonal/Monthly and Composite Airshed}
\usage{
airshed(loc, lon, lat, day_select=NA, month_select=NA, year_select,
   season_name="season",mode="seasonal", hy_alt, sPix=0.3,
   nPix=100, cutoff=10, composite=TRUE, ...)
}
\arguments{
\item{loc}{character. Name of location, starting point, used in naming subdirectories and files}

\item{lon}{numeric. Longitude of starting point}

\item{lat}{numeric. Latitude of starting point}

\item{day_select}{numeric vector, integers, [1,365]. Days to be run, in Julian days, non-leap years, useful if mode="seasonal"}

\item{month_select}{numeric vector, integers, [1,12]. Months to be run, useful if mode="monthly"}

\item{year_select}{numeric vector, integers. Years to be run}

\item{season_name}{character. Name of season, used in the names of the output folder and file [default is "season"]}

\item{mode}{character: "monthly" or "seasonal" [default is "seasonal"]}

\item{hy_alt}{character string or vector. Name(s) of hysplit heights to be run}

\item{sPix}{numeric. Size of pixel, in degrees [default is 0.3]}

\item{nPix}{numeric. Number of pixels in horizontal and vertical direction [default is 100]}

\item{cutoff}{numeric. Threshold, in percent, for pixels to be considered part of seasonal airshed  [default to 10]}

\item{composite}{logical. Create composite airshed? [default is TRUE]}

\item{alt.adj}{numeric. Adjustment if more than one location [default is 0]}

\item{...}{see global arguments: pointDir, ask_home}
}
\description{
Create monthly/seasonal and composite airsheds from HYSPLIT back trajectories for a given point location
}
\examples{
#creates winter delhi airshed for 2007-2013 and composite winter airshed for these years
airshed(loc="delhi",lon=77.2090,lat=28.6139,month_select=10:11,year_select=2007:2013,season_name="win",hy_alt="500m")
}
\keyword{airshed}
\keyword{composite}
\keyword{hysplit}
\keyword{monthly}
\keyword{seasonal}

