% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyHyFRP.R
\name{dailyHyFRP}
\alias{dailyHyFRP}
\title{Extract FRP from Daily HYSPLIT Airshed}
\usage{
dailyHyFRP(loc, day_select, year_select, buf=10000,
   hy_alt, nHours, local_time, frp_mode="combined", ...)
}
\arguments{
\item{loc}{character. Name of location, starting point, used in naming subdirectories and files}

\item{day_select}{numeric vector, integers, [1,365]. Days to be run, in julian days, non-leap years}

\item{year_select}{numeric vector, integers. Years to be run}

\item{buf}{numeric. Width of buffer, in meters, from original trajectory [default is 10000]}

\item{hy_alt}{character string or vector. Name(s) of hysplit heights to be run}

\item{nHours}{integer. Total hours in back trajectories.}

\item{local_time}{numeric. local time of HYSPLIT starting time, in hours}

\item{frp_mode}{character: "terra", "aqua", or "combined". Choose Terra-only, Aqua-only, or Terra+Aqua FRP. [default is "combined"]}

\item{...}{see global arguments: rasDir, pointDir, ask_home}
}
\description{
Extract the daily FRP from HYSPLIT-derived airsheds of selected point location(s)
}
\examples{
#extracts daily FRP in October and November, 2007-2013
#from a 10-km buffer (20 km in width) of 72-hour HYSPLIT
#back trajectories started from Delhi at 5:00pm (17:00) local time
dailyHyFRP(loc="delhi",day_select=274:334,year_select=2007:2013,buf=10000,hy_alt="500m",local_time=17)
}
\keyword{airshed}
\keyword{daily}
\keyword{frp}
\keyword{hysplit}
\keyword{modis}

