% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyProcess.R
\name{HyProcess}
\alias{HyProcess}
\title{Process HYSPLIT files and save coordinates to table}
\usage{
HyProcess(loc, year_select, hy_alt, nHours, alt.adj=0, ...)
}
\arguments{
\item{loc}{character. Name of location, starting point, used in naming subdirectories and files}

\item{year_select}{numeric vector. Years to be run}

\item{hy_alt}{character string or vector. Name(s) of hysplit heights to be run}

\item{nHours}{numeric. Total hours in back trajectories}

\item{alt.adj}{numeric. Adjustment if more than one location [default is 0]}

\item{...}{see global arguments: hyDir, pointDir, ask_home}
}
\description{
Process HYSPLIT files and save coordinates to table
}
\examples{
#Process all available HYSPLIT downloads and organizes the extracted coordinates of the back trajectories into tables 
HyProcess(loc="delhi",year_select=2007:2013,hy_alt="0.5km",nHours=72)
}
\keyword{hysplit}
\keyword{process}

