% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeFRPstack.R
\name{modeFRPstack}
\alias{modeFRPstack}
\title{Sum and Stack MODIS Daily FRP to Monthly or Seasonal}
\usage{
modeFRPstack(loc=NA, day_select=NA, month_select=NA,
   year_select, mode="seasonal", frp_mode="combined",
   season_name="season", next_yr=NA, ...)
}
\arguments{
\item{loc}{character. Name of location, starting point, used in naming subdirectories and files when mode="seasonal" [default is NA]}

\item{day_select}{numeric vector, integers, [1,365]. Days to be run, in Julian days, non-leap years, useful if mode="seasonal"}

\item{month_select}{numeric vector, integers, [1,12]. Months to be run, useful if mode=="monthly"}

\item{year_select}{numeric vector. Years to be run}

\item{mode}{character: "monthly" or "seasonal".  [default is "seasonal"]}

\item{frp_mode}{character: "terra", "aqua", or "combined". Choose Terra-only, Aqua-only, or Terra+Aqua FRP. [default is "combined"]}

\item{season_name}{character. Name of season, used in the names of the output subfolder and file [default is “season"]}

\item{next_yr}{numeric [1,365]. Days in next year, format is Julian days in non-leap year, useful for stacking days in winter months, only for "seasonal" mode [default is NA]}

\item{...}{see global arguments: rasDir, pointDir, ask_home}
}
\description{
Stack and sum daily FRP to monthly and seasonal, up to a two-year period (for HYSPLIT). Input takes dailyFRPstack() output.
}
\examples{
#stacks daily FRP to monthly for January-April from 2007-2013
modeFRPstack(year_select=2007:2013,month_select=1:4,mode="monthly")
 
#stacks daily FRP for the days 152-243 (leap:152-244) (JJA) in 2007-2013
modeFRPstack(year_select=2007:2013,mode="seasonal",day_select=152:243,season_name="summer")

#stacks daily FRP for the days 335-365 (leap:336-366) plus days 1-59 (leap:1-60) of the next year,
or DJF, in 2007-2013 (requires January FRP in 2014)
modeFRPstack(year_select=2007:2013,mode="seasonal",day_select=274:300,season_name="winter",next_yr=1:31)

#stacks daily FRP for Feburary 25-29, 2008
#in leap years, a day_select ending on day 59
#has day 60 added to the stack
modeFRPstack(year_select=2008,mode="seasonal",day_select=56:59,season_name="EndofFeb")
}
\keyword{frp}
\keyword{modis}
\keyword{monthly}
\keyword{seasonal}
\keyword{stack}

