% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanMask.R
\name{oceanMask}
\alias{oceanMask}
\title{Create an Ocean Mask}
\usage{
oceanMask(...)
}
\arguments{
\item{...}{see global arguments: ask_home}
}
\description{
Create an ocean mask from the countries.tif raster derived from the
Natural Earth shapefile ne_10m_admin_0_countries using QGIS.
}
\keyword{mask}
\keyword{ocean}

