% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.airshedC.R
\name{plot.airshedC}
\alias{plot.airshedC}
\title{Create Plots for HYSPLIT-derived Composite Seasonal Airshed}
\usage{
plot.airshedC(loc, lon, lat, season_name, borderName="ne_10m_admin_0_countries",
   x.adj=FALSE, zoom=4.5, x.just=0, y.just=0, nKM=150, col.adj=c(0,0,0,0),
   scale.adj=c(0,0), caption=NA ,caption.adj=c(0,0),
   nYrs.adj=c(0,0), legend.adj=c(0,0), txt_size=0.75, ...)
}
\arguments{
\item{loc}{character. Name of location, starting point, used in naming subdirectories and files}

\item{lon}{numeric. Longitude of starting point}

\item{lat}{numeric. Latitude of starting point}

\item{season_name}{character. Name of season, used in the names of the output folder and file}

\item{borderName}{character. Name of shapefile located in the "shapefiles" folder of project home directory to plot country/sub-country borders [default is "ne_10m_admin_0_countries"]}

\item{x.adj}{logical. Adjust x axis latitude to 0-360? [default is FALSE] (-180 to 180)}

\item{zoom}{numeric. Window of the plot, in degrees, from the lon, lat of the starting location in all 4 directions [default is 4.5]}

\item{x.just}{numeric. Adjust window laterally, positive is right and negative is left [default is 0]}

\item{y.just}{numeric. Adjust window vertically, positive is right and negative is left [default is 0]}

\item{nKM}{numeric. Length of scale bar, in kilometers [default is 150]}

\item{col.adj}{numeric vector, length=4. Adjust color bar position/size (format is c(x.min, x.max, y.min, y.max) adjustment) [default is c(0,0,0,0)]}

\item{scale.adj}{numeric vector, length=2. Adjust scale bar position (format is c(x, y) adjustment), [default is c(0,0)]}

\item{caption}{character or expression. Name of caption, [default is NA], which is no caption}

\item{caption.adj}{numeric vector, length=2. Adjust caption position (format is c(x, y) adjustment) [default is c(0,0)]}

\item{nYrs.adj}{numeric vector, length=2. Adjust number of years label position (format is c(x, y) adjustment) [default is c(0,0)]}

\item{legend.adj}{numeric vector, length=2. Adjust legend position (format is c(x, y) adjustment), [default is c(0,0)]}

\item{txt_size}{numeric. Adjust text size of plot labels, based on size of axis tick labels [default is 0.75]}

\item{...}{see global arguments: pointDir, ask_home}
}
\description{
Create plot for composite seasonal airshed for a given point location
}
\examples{
#creates plot for composite winter delhi airshed for 2007-2013
plot.airshedC(loc="delhi",lon=77.2090,lat=28.6139,x.just=2,y.just=-2,caption="Delhi Winter Composite Airshed \\nOct-Nov")
}
\keyword{airshed}
\keyword{hysplit}
\keyword{plot}
\keyword{seasonal}

