% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.airshedM.R
\name{plot.airshedM}
\alias{plot.airshedM}
\title{Create Plots for HYSPLIT-derived Monthly Airshed}
\usage{
plot.airshedM(loc, lon, lat, month_select, year_select,
   tYearss=year_select[1]:year_select[length(year_select)],
   hy_alt, alt.adj=0, sPix=0.25, nPix=100, borderName="ne_10m_admin_0_countries",
   x.adj=FALSE, zoom=4.5, x.just=0, y.just=0, nKM=150,
   traj.col=c("#a6cee3","#fb9a99","#cab2d6"), traj.lwd=0.25,
   col.adj=c(0,0,0,0), scale.adj=c(0,0), bar.adj=c(0,0,0,0),
   bar2.adj=c(0,0,0,0), caption=NA, caption.adj=c(0,0),
   nAlt.adj=c(0,0), legend.lab=hy_alt, legend.adj=c(0,0),
   txt_size=0.75, ...)
}
\arguments{
\item{loc}{character. Name of location, starting point, used in naming subdirectories and files}

\item{lon}{numeric. Longitude of starting point}

\item{lat}{numeric. Latitude of starting point}

\item{month_select}{numeric vector. Months to be run}

\item{year_select}{numeric vector. Years to be run}

\item{hy_alt}{character string or vector. Name(s) of hysplit heights to be run}

\item{alt.adj}{numeric. Adjustment if more than one location [default is 0]}

\item{sPix}{numeric. Size of pixel, in degrees [default is 0.3]}

\item{nPix}{numeric. Number of pixels in horizontal and vertical direction [default is 100]}

\item{borderName}{character. Name of shapefile located in the "shapefiles" folder of project home directory to plot country/sub-country borders [default is "ne_10m_admin_0_countries"]}

\item{x.adj}{logical. Adjust x axis latitude to 0-360? default is FALSE (-180 to 180)}

\item{zoom}{numeric. Window of the plot, in degrees, from the lon, lat of the starting location in all 4 directions [default is 4.5]}

\item{x.just}{numeric. Adjust window laterally, positive is right and negative is left [default is 0]}

\item{y.just}{numeric. Adjust window vertically, positive is right and negative is left [default is 0]}

\item{nKM}{numeric. Length of scale bar, in kilometers [default is 150000]}

\item{traj.col}{character string or vector in html or R color name. Colors of trajectories [default is c("#a6cee3","#fb9a99","#cab2d6")]}

\item{traj.lwd}{numeric. Width of trajectory lines [default is 0.3]}

\item{col.adj}{numeric vector, length=4. Adjust color bar position/size (format is c(x.min, x.max, y.min, y.max) adjustment) [default is c(0,0,0,0)]}

\item{scale.adj}{numeric vector, length=2. Adjust scale bar position (format is c(x, y) adjustment), [default is c(0,0)]}

\item{bar.adj}{numeric vector, length=4. Adjust timestamp bar position/size (format is c(x.min, x.max, y.min, y.max) adjustment), [default is c(0,0,0,0)]}

\item{caption}{character or expression. Name of caption, [default is NA], which is no caption}

\item{caption.adj}{numeric vector, length=2. Adjust caption position (format is c(x, y) adjustment) [default is c(0,0)]}

\item{nAlt.adj}{numeric vector, length=2. Adjust number of trajectories label position (format is c(x, y) adjustment) [default is c(0,0)]}

\item{legend.lab}{character vector. Labels for back trajectories [default is hy_alt]}

\item{legend.adj}{numeric vector, length=2. Adjust legend position (format is c(x, y) adjustment), [default is c(0,0)]}

\item{txt_size}{numeric. Adjust text size of plot labels, based on size of axis tick labels [default is 0.75]}

\item{tYearss}{numeric vector. Total range of years in study, ignoring missing years [default = year_select[1]:year_select[length(year_select)], which is the years from the first year to last year in year_select}

\item{...}{see global arguments: pointDir, ask_home}
}
\description{
Create plots for monthly airsheds with daily HYSPLIT back trajectories for a given point location
}
\examples{
#creates plots for monthly delhi airshed for Oct-Nov, 2007-2013
plot.airshedM(loc="delhi",lon=77.2090,lat=28.6139,month_select=10:11,year_select=2007:2013,hy_alt="0.5 km",x.just=1.5,y.just=-1.5,caption="Delhi Monthly Airshed")
}
\keyword{airshed}
\keyword{hysplit}
\keyword{monthly}
\keyword{plot}

